Update backEnd/src/main/java/org/infinity/sixtalebackend/global/config/WebConfig.java, backEnd/src/main/java/org/infinity/sixtalebackend/infra/websocket/config/WebSocketConfig.java
